Use the targetDetent when updating detents if the sheet is not idle
Modal sheet uses a Hidden bottom sheet and sets its detent to FullyExpanded, once the modal has been introduced to the screen. In the case where the sheet's detents need to be updated (ie due to a resize) we used to set the previously set currentDetent, which in the case of ModalBottomSheets this is most likely Hidden. The fix is to use the targetDetent as the new detent when updating the detents, if the sheet is not idle.

Fixes https://github.com/composablehorizons/composables-core/issues/18

(cherry picked from commit f14cba4d9dc43e447ef84aeb86ce3811f4feb916)