Replace usage of KeyDownHandler with Compose onKeyEvent
The KeyDownHandler was using each platform's native key handling mechanism, which doesn't play nice with Compose's key events system and it caused bugs. One example would be trying to hide a Menu on JVM by pressing Esc. The Menu could not consume the Compose key event, as a result other onKeyEvent listeners in the tree are triggered.