Remove capping of sheet content height
This fixes a bug where the sheet height is say 100.dp, and the detent is 50.dp. Visually, we want only the top part of the sheet to be visible. However, forcing a height might shrink the contents of the sheet in the 50.dp height if the contents use say `Modifier.weight()`.

This block used to be necessary when the sheet was first introduced, but doing a few test it is no longer seem to be needed.